Tracking allocations in the interpreter
I tried to document all possible allocations in the interpreter. This
way we should cover everything except builtin functions (which need
manual review).

While doing shuch a task, it is clear that it is impossible to really
track what is happening because of escape analysis and/or
devirtualization. For this reason I applyed these rough simplifications:

 - all transformations of a value to an interface will be counted as one
   allocation (e.g. disregarding devirtualization, small value
   optimization, zero value optimization, etc...);
 - an interface counts as 1 allocation;
 - strings are always allocated in the heap;
 - string slices never allocate;
 - all slices with variable lenght are escaped to the heap.

While not true in general, I think they greatly simplify the reasoning
while estimating the amount of memory used.