Memory bounding default behaviour (#7)
Due to an oversight, the allocation tracking functions added in #4 require that the user SetMaxAllocs before any allocs are declared. This will break applications which do not care about allocs once we add alloc-declarations to the standard library.

This PR makes the thread consider a max alloc limit of 0 to be unbound.