Approximate map measurements
Maps are hard to measure since they lack any way to get the "capacity",
whatever that mean. Other languages "sloved" the problem by letting you
access the count of buckets, but this led to the unwanted result of
forcing bucketed hash maps (which is bad).

After a bit of digging we decided to take the approximation/experimental
route. We collected some data on memory usage and approximated map usage
with the only discoverable measure: the length.

The constant still come from some implementation details (of course),
but we made sure to be pessimistic enough. One thing to bear in mind is
that once fully grown, go maps never shrink, so as pessimistic one can
be, there will always be cases where many additions followed by many
deletions will make the map very big and we will not be able to take
that into account.