Make SafeStringBuilder update Thread allocation count (#110)
The existing SafeStringBuilder implementation uses CheckAllocs when resizing the buffer. This has three problems:

1. If other allocations are made as the string builder is being used, the thread.Allocs() count doesn’t represent the true total number of allocations
2. It is easy for a user to forget to count the allocations of the built string after they’ve finished using the safe string builder
3. The use of CheckAllocs while running is inconsistent with SafeAppender, which uses AddAllocs

The use of CheckAllocs rather than AddAllocs was born out of a need to use SafeStringBuilder in the print and fail builtins, where the built string must be bounded, but ultimately its ownership passes out of the starlark environment (e.g. to a log).

With this change, allocations are actually tracked using AddAllocs, and a new Allocs method is added to support cases where the string is knowingly being deallocated.