Fixing allocation pattern for SafeStringBuilder (#107)
It seems that the `Grow` function in Go doesn't uses all the bytes
available coming from a reallocation. This would force us to compute
correctly the size of an allocation after each `SafeStringBuilder`
usage.

Given that the objective is to make it as simple as possible for the end
user to use these tools, this change removes that need.